+++ backend-starter-spring/src/main/kotlin/dev/syncforge/backendstarterspring/web/WebConfiguration.kt
@@ -1,0 +1,12 @@
+package dev.syncforge.backendstarterspring.web
+
+import org.springframework.context.annotation.Configuration
+import org.springframework.http.converter.HttpMessageConverter
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer
+
+@Configuration
+class WebConfiguration : WebMvcConfigurer {
+    override fun extendMessageConverters(converters: MutableList<HttpMessageConverter<*>>) {
+        converters.add(0, SyncForgeJsonHttpMessageConverter())
+    }
+}